fix "com.t1m0.logging" module definition
If a module needs some other classes, it is typically a better idea
to add a module dependency. Adding another artifact to the module
can lead to the same class being loaded by multiple classloaders
(one classloader for each module), which leads to different `static`
variables, higher memory usage, and general unhappiness.

Also modules can have their own dependencies. For example,
JBoss LogManager requires WildFly Common and if it isn't present,
classloading errors will pop up. If we added the JBoss LogManager
artifact to the `com.t1m0.logging` module, we'd also have to add
the WildFly Common artifact. However, if we add _a dependency_
on the JBoss LogManager module, we don't have to care, because
the module takes care of that itself.