remove dependency on "com.t1m0.logging:logging" from the app
The application had a direct dependency on the `com.t1m0.logging:logging`
artifact, which isn't necessary. This artifact is present in the module
defined in `src/main/resources/modules`, so it doesn't have to be
present in the application again.

Moreover, depending on this artifact would bring its dependencies
directly into the application, which can lead to classloading issues.
For example, if we had a dependency on this artifact, the application
would include its own copy of the JBoss LogManager classes, and if
we used the JBoss LogManager classes in the app (e.g. if we accessed
the JBoss LogManager `MDC` class, like `MDCWrapper` used to), we would
get a different class then the rest of the application server.

This commit also makes the dependency on SFL4J API `<scope>provided</scope>`,
because the application server will provide the SLF4J API classes to the app.