add an example "jboss-deployment-structure.xml" file
If the application had to use JBoss LogManager classes directly
(it doesn't now, since `MDCWrapper` is gone), then a file like
this would have to be present. Presence of this file would then
make sure that the application server makes the JBoss LogManager
classes available to the deployment. Without that, accessing
the JBoss LogManager classes in the application results in
classloading issues.